+++ src/enum/InterceptEvent.ts
@@ -2,7 +2,8 @@ enum InterceptEvent {
     PRE_REQUEST,
     POST_REQUEST,
     PRE_RESPONSE,
-    POST_RESPONSE
+    POST_RESPONSE,
+    ERROR
 }
 
 export default InterceptEvent;