+++ src/type/GetRequestConfig.ts
@@ -3,6 +3,7 @@ import StandardObjectType from "@src/type/StandardObjectType";
 type GetRequestConfig = {
     url: string;
     query?: StandardObjectType | URLSearchParams;
+    signal?: AbortSignal
 }
 
 export default GetRequestConfig;